+++ baipw/tests/test_response.py
@@ -1,3 +1,4 @@
+import django
 from django.test import RequestFactory, TestCase, override_settings
 
 from baipw.response import DEFAULT_AUTH_TEMPLATE, HttpUnauthorizedResponse
@@ -34,3 +35,17 @@ class TestHttpUnauthorizedResponse(TestCase):
             response.get_response_content().strip(), "This is a test template."
         )
         self.assertEqual(response["Content-Type"], "text/html")
+
+    def test_never_cache_headers_set(self):
+        response = HttpUnauthorizedResponse(request=self.request)
+        cache_control_args = [v.strip() for v in response["cache-control"].split(",")]
+        if django.VERSION >= (3, 0):
+            self.assertCountEqual(
+                cache_control_args,
+                ["max-age=0", "no-cache", "no-store", "must-revalidate", "private"],
+            )
+        else:
+            self.assertCountEqual(
+                cache_control_args,
+                ["max-age=0", "no-cache", "no-store", "must-revalidate"],
+            )